Update "Build & Test" Action

+++ performance-tests/bench_2/node_controller/ProcessStatsCollector.h
@@ -24,8 +24,8 @@ private:
 #ifndef ACE_HAS_MAC_OSX
   int process_id_;
   size_t num_processors_;
-  size_t total_virtual_mem_;
-  size_t total_mem_;
+  uint64_t total_virtual_mem_;
+  uint64_t total_mem_;
 #endif
 #ifdef ACE_WIN32
   HANDLE process_handle_;
